+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Ember Flexberry Changelog
 ### New Features & improvements
 
+## 2016-06-21
+### Breaking changes
+FlexberryObjectlistview:
+* Change of mechanism for adding custom user buttons.
+
 ## 2016-06-14
 ### Fixed
 FlexberryTextboxComponent, FlexberryTextareaComponent: